AstronomicalTime.Epoch — Method.

Epoch{T}(year, month, day,
    hour=0, minute=0, seconds=0, milliseconds=0) where T<:Timescale

Construct an Epoch with timescale T at the given date and time.

Example

julia> Epoch{TT}(2017, 3, 14, 7, 18, 20, 325)
2017-03-14T07:18:20.325 TT

source

AstronomicalTime.Epoch — Method.

Epoch{T}(timestamp::AbstractString) where T<:Timescale

Construct an Epoch with timescale T from a timestamp.

Example

julia> Epoch{TT}("2017-03-14T07:18:20.325")
2017-03-14T07:18:20.325 TT

source

AstronomicalTime.Epoch — Method.

Epoch{T}(ep::Epoch{S}) where {T<:Timescale, S<:Timescale}

Convert an Epoch with timescale S to an Epoch with timescale T.

Example

julia> Epoch{TT}(Epoch{TAI}(2000, 1, 1))
2000-01-01T00:00:32.184 TT

source

AstronomicalTime.Epoch — Method.

Epoch{T}(dt::DateTime) where T<:Timescale

Convert a DateTime object to an Epoch with timescale T.

Example

julia> Epoch{TT}(DateTime(2017, 3, 14, 7, 18, 20, 325))
2017-03-14T07:18:20.325 TT

source

AstronomicalTime.Epoch — Method.

Epoch{T}(jd1, jd2=0.0) where T<:Timescale

Construct an Epoch with timescale T from a two-part Julian date.

Example

julia> Epoch{TT}(2.4578265e6, 0.30440190993249416)
2017-03-14T07:18:20.325 TT

source

AstronomicalTime.Timescale — Type.

All timescales are subtypes of the abstract type Timescale. The following timescales are defined:

UTC — Coordinated Universal Time

UT1 — Universal Time

TAI — International Atomic Time

TT — Terrestrial Time

TCG — Geocentric Coordinate Time

TCB — Barycentric Coordinate Time

TDB — Barycentric Dynamical Time

source

AstronomicalTime.@timescale — Macro.

@timescale scale

Define a new timescale and the corresponding Epoch type alias.

Example

julia> @timescale Custom

julia> Custom <: Timescale
true
julia> CustomEpoch == Epoch{Custom}
true

source

Base.Dates.DateTime — Method.

DateTime{T<:Timescale}(ep::Epoch{T})

Convert an Epoch with timescale T to a DateTime object.

Example

julia> DateTime(Epoch{TT}(2017, 3, 14, 7, 18, 20, 325))
2017-03-14T07:18:20.325

source