AstronomicalTime.Epoch — Type.

Epoch{T}(timestamp::AbstractString) where T<:Timescale

Construct an Epoch with timescale T from a timestamp.

Example

julia> Epoch{TT}("2017-03-14T07:18:20.325")
2017-03-14T07:18:20.325 TT

source

AstronomicalTime.Timescale — Type.

All timescales are subtypes of the abstract type Timescale. The following timescales are defined:

UTC — Coordinated Universal Time

UT1 — Universal Time

TAI — International Atomic Time

TT — Terrestrial Time

TCG — Geocentric Coordinate Time

TCB — Barycentric Coordinate Time

TDB — Barycentric Dynamical Time

source

AstronomicalTime.@timescale — Macro.

@timescale scale

Define a new timescale and the corresponding Epoch type alias.

Example

julia> @timescale Custom

julia> Custom <: Timescale
true
julia> CustomEpoch == Epoch{Custom}
true

source

Base.Dates.DateTime — Method.

DateTime{T<:Timescale}(ep::Epoch{T})

Convert an Epoch with timescale T to a DateTime object.

Example

julia> DateTime(Epoch{TT}(2017, 3, 14, 7, 18, 20, 325))
2017-03-14T07:18:20.325

source